dev(python controller client): update lib usage
