+++ clients/python/inspr/__init__.py
@@ -1,6 +1,3 @@
 from .client import *
 from .rest import *
-from .controller.app import *
-from .controller.channel import *
-from .controller.type import *
-from .controller.alias import *
+from .controller.controller_client import *
